@@ -1,90 +1,29 @@
 {
-    "name": "hade",
-    "version": "1.0.0",
-    "description": "A Vue.js project",
-    "author": "yejianfeng <jianfengye110@gmail.com>",
-    "private": true,
-    "scripts": {
-        "dev": "webpack-dev-server --inline --progress --config build/webpack.dev.conf.js",
-        "start": "npm run dev",
-        "unit": "jest --config test/unit/jest.conf.js --coverage",
-        "e2e": "node test/e2e/runner.js",
-        "test": "npm run unit && npm run e2e",
-        "lint": "eslint --ext .js,.vue src test/unit test/e2e/specs",
-        "build": "node build/build.js",
-        "docs:dev": "vuepress dev docs",
-        "docs:build": "vuepress build docs"
-    },
-    "dependencies": {
-        "vue": "^2.5.2",
-        "vue-router": "^3.0.1"
-    },
-    "devDependencies": {
-        "autoprefixer": "^7.1.2",
-        "babel-core": "^6.22.1",
-        "babel-eslint": "^8.2.1",
-        "babel-helper-vue-jsx-merge-props": "^2.0.3",
-        "babel-jest": "^21.0.2",
-        "babel-loader": "^7.1.1",
-        "babel-plugin-dynamic-import-node": "^1.2.0",
-        "babel-plugin-syntax-jsx": "^6.18.0",
-        "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
-        "babel-plugin-transform-runtime": "^6.22.0",
-        "babel-plugin-transform-vue-jsx": "^3.5.0",
-        "babel-preset-env": "^1.3.2",
-        "babel-preset-stage-2": "^6.22.0",
-        "babel-register": "^6.22.0",
-        "chalk": "^2.0.1",
-        "chromedriver": "^2.27.2",
-        "copy-webpack-plugin": "^4.0.1",
-        "cross-spawn": "^5.0.1",
-        "css-loader": "^0.28.0",
-        "eslint": "^4.15.0",
-        "eslint-config-standard": "^10.2.1",
-        "eslint-friendly-formatter": "^3.0.0",
-        "eslint-loader": "^1.7.1",
-        "eslint-plugin-import": "^2.7.0",
-        "eslint-plugin-node": "^5.2.0",
-        "eslint-plugin-promise": "^3.4.0",
-        "eslint-plugin-standard": "^3.0.1",
-        "eslint-plugin-vue": "^4.0.0",
-        "extract-text-webpack-plugin": "^3.0.0",
-        "file-loader": "^1.1.4",
-        "friendly-errors-webpack-plugin": "^1.6.1",
-        "html-webpack-plugin": "^2.30.1",
-        "jest": "^22.0.4",
-        "jest-serializer-vue": "^0.3.0",
-        "nightwatch": "^0.9.12",
-        "node-notifier": "^5.1.2",
-        "optimize-css-assets-webpack-plugin": "^3.2.0",
-        "ora": "^1.2.0",
-        "portfinder": "^1.0.13",
-        "postcss-import": "^11.0.0",
-        "postcss-loader": "^2.0.8",
-        "postcss-url": "^7.2.1",
-        "rimraf": "^2.6.0",
-        "selenium-server": "^3.0.1",
-        "semver": "^5.3.0",
-        "shelljs": "^0.7.6",
-        "uglifyjs-webpack-plugin": "^1.1.1",
-        "url-loader": "^0.5.8",
-        "vue-jest": "^1.0.2",
-        "vue-loader": "^13.3.0",
-        "vue-style-loader": "^3.0.1",
-        "vue-template-compiler": "^2.5.2",
-        "vuepress": "^1.8.2",
-        "webpack": "^4.8.3",
-        "webpack-bundle-analyzer": "^2.9.0",
-        "webpack-dev-server": "^2.9.1",
-        "webpack-merge": "^4.1.0"
-    },
-    "engines": {
-        "node": ">= 6.0.0",
-        "npm": ">= 3.0.0"
-    },
-    "browserslist": [
-        "> 1%",
-        "last 2 versions",
-        "not ie <= 8"
-    ]
+  "name": "element-plus-vite-starter",
+  "private": true,
+  "version": "0.1.0",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "generate": "vite-ssg build",
+    "preview": "vite preview",
+    "typecheck": "vue-tsc --noEmit"
+  },
+  "dependencies": {
+    "element-plus": "^2.3.6",
+    "vue": "^3.3.4"
+  },
+  "devDependencies": {
+    "@iconify-json/ep": "^1.1.10",
+    "@types/node": "^20.3.1",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "sass": "^1.63.4",
+    "typescript": "^5.1.3",
+    "unocss": "^0.53.1",
+    "unplugin-vue-components": "^0.25.1",
+    "vite": "^4.3.9",
+    "vite-ssg": "^0.22.2",
+    "vue-tsc": "^1.8.0"
+  },
+  "license": "MIT"
 }
